closes #34:extracted utility functions from events
